Control.Effect.Error

Documentation

class Monad m => Error e m where

Error e is an effect that allows throwing and catching errors of type e. Note that these have no relation to Haskell’s built-in support for synchronous or asynchronous runtime exceptions provided from Control.Exception.

Instances should obey the law catch (throw x) f ≡ pure (f x).

Methods

throw :: e -> m a

Raises an error of type e.

catch :: m a -> (e -> m a) -> m a

Runs the given sub-computation. If it raises an error of type e, the error is provided to the given handler function, and execution resumes from the point of the call to catch.

Pure Error handler

For historical reasons, the canonical handler for the Error effect provided by transformers has the name ExceptT, not ErrorT. To avoid confusion, this library does not export it under a different alias. However, the handler function is still named runError.

runError :: EffT (ExceptT e) m a -> m (Either e a)

Handles an Error e effect. Returns Left if the computation raised an uncaught error, otherwise returns Right.

data ExceptT e (m :: Type -> Type) a

A monad transformer that adds exceptions to other monads.

ExceptT constructs a monad parameterized over two things:

e - The exception type.

m - The inner monad.

The return function yields a computation that produces the given value, while >>= sequences two subcomputations, exiting on the first exception.