Control.Effect.Internal.SmallArray

Documentation

data SmallArray a

Constructors

SmallArray (SmallArray# a)

data SmallMutableArray s a

Constructors

SmallMutableArray (SmallMutableArray# s a)

newSmallArray :: (DebugCallStack, PrimMonad m) => Int -> a -> m (SmallMutableArray (PrimState m) a)

runSmallArray :: (forall s. ST s (SmallMutableArray s a)) -> SmallArray a

unsafeFreezeSmallArray :: PrimMonad m => SmallMutableArray (PrimState m) a -> m (SmallArray a)

Render a mutable array immutable.

This operation performs no copying, so care must be taken not to modify the input array after freezing.

unsafeThawSmallArray :: PrimMonad m => SmallArray a -> m (SmallMutableArray (PrimState m) a)

Render an immutable array mutable.

This operation performs no copying, so care must be taken with its use.

sizeofSmallArray :: DebugCallStack => SmallArray a -> Int

sizeofSmallMutableArray :: DebugCallStack => SmallMutableArray s a -> Int

indexSmallArray :: DebugCallStack => SmallArray a -> Int -> (# a #)

readSmallArray :: (DebugCallStack, PrimMonad m) => SmallMutableArray (PrimState m) a -> Int -> m a

writeSmallArray :: (DebugCallStack, PrimMonad m) => SmallMutableArray (PrimState m) a -> Int -> a -> m ()

copySmallArray :: (DebugCallStack, PrimMonad m) => SmallMutableArray (PrimState m) a -> Int -> SmallArray a -> Int -> Int -> m ()

cloneSmallArray :: DebugCallStack => SmallArray a -> Int -> Int -> SmallArray a

copySmallMutableArray :: (DebugCallStack, PrimMonad m) => SmallMutableArray (PrimState m) a -> Int -> SmallMutableArray (PrimState m) a -> Int -> Int -> m ()

cloneSmallMutableArray :: (DebugCallStack, PrimMonad m) => SmallMutableArray (PrimState m) a -> Int -> Int -> m (SmallMutableArray (PrimState m) a)