Control.Effect.Reader

Documentation

class Monad m => Reader r m where

Reader r is an effect that provides access to a global environment of type r.

Instances should obey the law f <$> ask ≡ asks f ≡ local f ask.

Minimal complete definition

(ask | asks), local

Methods

ask :: m r

Retrieves a value from the environment.

asks :: (r -> a) -> m a

Applies a function to a value in the environment and returns the result.

local :: (r -> r) -> m a -> m a

Runs a subcomputation in an environment modified by the given function.

Pure Reader handler

runReader :: r -> EffT (ReaderT r) m a -> m a

Handles a Reader r effect by supplying a value for the environment.

data ReaderT r (m :: Type -> Type) a

The reader monad transformer, which adds a read-only environment to the given monad.

The return function ignores the environment, while >>= passes the inherited environment to both subcomputations.