Control.Effect.Resource

Description

This module provides a set of classes and operations for safe resource allocation and disposal.

Masking interrupts

class Monad m => Mask m where

An effect that provides the ability to temporarily mask asynchronous interrupts.

Unlike most effects, this effect does not have its own handler; there is no separate runMask operation. This is because the usual handler for Mask is IO, which is always at the base of a monad transformer stack.

Mask may be used in pure computations (where both mask and uninterruptibleMask are no-ops), but it can only be discharged by runIdentity, which serves as a proof that the computation is, in fact, pure.

Methods

mask :: ((forall a. m a -> m a) -> m b) -> m b

Executes a computation with asynchronous interrupts masked; see Control.Exception.mask for details.

uninterruptibleMask :: ((forall a. m a -> m a) -> m b) -> m b

Executes a computation with asynchronous interrupts uninterruptibly masked. This should be used with great care; see Control.Exception.uninterruptibleMask for details.

mask_ :: Mask m => m a -> m a

Like mask, but does not pass a restore action to its argument.

uninterruptibleMask_ :: Mask m => m a -> m a

Like uninterruptibleMask, but does not pass a restore action to its argument.

Registering cleanup actions

class Monad m => MonadUnwind m where

The class of monads that support registering cleanup actions on failure. Note that this is not an effect; it cannot be automatically lifted through EffT the way effects can. Doing so would be unsafe, as different handlers may provide arbitrarily different notions of failure, and they must supply MonadUnwind instances to ensure those failures are not ignored. However, note that any handlers defined using HandlerT do not need their own MonadUnwind instances, as they will inherit the instance of their underlying handlers.

Methods

onException :: m a -> m b -> m a

(a `onException` b) runs a. If and only if it fails with an error, b is executed for side effects (with asynchronous exceptions masked, if relevant), after which the error is re-raised. If b fails with an error, its error takes priority over the error raised by a.

onException cannot generally be used to ensure the disposal of an acquired resource because an asynchronous exception might be raised after the resource is acquired but before the exception handler is installed. For safe resource management, use bracket or bracketOnError_ instead.

Note that because the error is re-raised after b is executed, changes to the monadic state made by b will be discarded for any effects handled more locally than the effect that triggered the failure.

Safe resource allocation and disposal

class (Mask m, MonadUnwind m) => Resource m

(Resource m) is a constraint synonym for (Mask m, MonadUnwind m), which together provide the operations necessary for safe resource allocation and disposal. The most frequently used operations of Resource are bracket and bracketOnError.

finally :: Resource m => m a -> m b -> m a

Like onException, but the action provided for the second argument is unconditionally executed, whether an error was raised or not.

bracket

Arguments

:: Resource m

=> m a
Acquire some resource.

-> (a -> m c)
Dispose the resource.

-> (a -> m b)
Use the resource.

-> m b

Safely acquires a resource for the dynamic extent of a nested computation given actions to acquire and dispose it. See also Control.Exception.bracket.

bracket_ :: Resource m => m a -> m c -> m b -> m b

Like bracket, but the return value from the first argument is ignored.

bracketOnError

Arguments

:: Resource m

=> m a
Acquire some resource.

-> (a -> m c)
Dispose the resource.

-> (a -> m b)
Use the resource.

-> m b

Like bracket, but the dispose action is only executed if an error is raised during the dynamic extent of the primary computation.

bracketOnError_ :: Resource m => m a -> m c -> m b -> m b

Like bracketOnError, but the return value from the first argument is ignored.