Control.Effect.State

Documentation

class Monad m => State s m where

State s is an effect that provides access to a single cell of mutable state of type s.

Minimal complete definition

get, (put | modify)

Methods

get :: m s

Retrieves the current value of the state.

put :: s -> m ()

Replaces the current state with the given value.

modify :: (s -> s) -> m ()

Modifies the current state by applying the given function to it.

Pure, strict State handler

Note: this handler provides put and modify operations that are strict in the state, even though put and modify from Control.Monad.Trans.State.Strict are lazy in the state. If you have a use case for lazy put/modify with a strict State handler, let me know, but in my experience this is what everyone wants, anyway.

runState :: Functor m => s -> EffT (StateT s) m a -> m (s, a)

evalState :: Monad m => s -> EffT (StateT s) m a -> m a

execState :: Monad m => s -> EffT (StateT s) m a -> m s

data StateT s (m :: Type -> Type) a

A state transformer monad parameterized by:

s - The state.

m - The inner monad.

The return function leaves the state unchanged, while >>= uses the final state of the first computation as the initial state of the second.