Control.Effect.State.Strict

Documentation

module Control.Effect.State

runState :: s -> Eff (State s ': effs) a -> Eff effs (s, a)

Handles a State effect using a strict cell of mutable state—each use of put or modify eagerly forces the new value. The state is initialized to the given value, and the final state is returned alongside the computation’s result.

evalState :: s -> Eff (State s ': effs) a -> Eff effs a

execState :: s -> Eff (State s ': effs) a -> Eff effs s