Control.Effect.Writer

Documentation

class (Monoid w, Monad m) => Writer w m where

Writer w is an effect that allows the accumulation of monoidal values of type w.

Instances should obey the following laws:

tell x *> tell y ≡ tell (x <> y)

listen (tell x) ≡ (x,) <$> tell x

censor f (tell x) ≡ tell (f x)

Methods

tell :: w -> m ()

Appends the given value to the current output.

listen :: m a -> m (w, a)

Executes the given action and includes its output in the result.

censor :: (w -> w) -> m a -> m a

Executes the given action and modifies its output by applying the given function.

Pure, strict Writer handler

Note: this handler is strict in the monoidal state to avoid space leaks, but this requires tell to produce left-associated uses of <>, which can be catastrophically slow on certain monoids, most notably []. To avoid pathological performance, use a data structure that supports efficient appends, such as Data.Sequence.Seq, or use Dual to flip the argument order of <> (but beware that this will cause the elements to be accumulated in reverse order).

runWriter :: (Monoid w, Functor m) => EffT (WriterT w) m a -> m (w, a)

evalWriter :: (Monoid w, Monad m) => EffT (WriterT w) m a -> m a

execWriter :: (Monoid w, Monad m) => EffT (WriterT w) m a -> m w

type WriterT w = HandlerT WriterH '[StateT w]

This handler is implemented on top of StateT rather than using a WriterT implementation from transformers because both the strict and lazy WriterT transformers leak space, and Control.Monad.Trans.Writer.CPS does not expose the WriterT constructor, precluding an efficient implementation of MonadTransControl.