Control.Effect.Internal

Documentation

type HandlerK = (Type -> Type) -> Type -> Type

The kind of effect handlers, which are monad transformers.

type EffectK = (Type -> Type) -> Constraint

The kind of effects, which are classes of monads.

newtype EffT (t :: HandlerK) m a

A monad transformer for handling effects. (EffT t) is actually no different from t at runtime, but it provides a different set of instances. Wrapping a monad transformer with EffT allows other effects to be automatically lifted through it, provided the underlying transformer provides a MonadTransControl instance.

EffT cannot be used with any arbitrary monad transformer: all monad transformers wrapped with EffT must provide an instance of the Handles type family to cooperate with effect with the effect lifting machinery. However, note that this requirement only applies to transformers wrapped in EffT directly, i.e. used as the t argument in (EffT t); there are no restrictions placed on the underlying monad (though effects will not be able to be automatically lifted through non-EffT layers).

Constructors

EffT

Fields

runEffT :: t m a

type family EffsT ts m where ...

A type alias for a stack of nested EffT transformers: EffsT '[t1, t2, ..., tn] m is equivalent to EffT t1 (EffT t2 (... (EffT tn m) ...)).

This can be considered the implementation dual to the Can interface.

Equations

EffsT '[] m = m

EffsT (t ': ts) m = EffT t (EffsT ts m)

newtype HandlerT tag ts m a

A helper for defining effect handlers in terms of other, existing handlers. (HandlerT tag ts) is equivalent to (EffsT ts), but the phantom tag parameter is useful as a way to disambiguate between different handler instances.

Constructors

HandlerT

Fields

runHandlerT :: EffsT ts m a

class Monad (EffsT ts m) => MonadTransHandler tag ts m where

Methods

liftHandler :: m a -> HandlerT tag ts m a

class LowerHandler tag t ts m where

Methods

lowerHandlerWith :: Run (EffT t) -> Run (HandlerT tag ts) -> HandlerT tag (t ': ts) m a -> m (StEffsT (t ': ts) a)

class MonadTransHandler tag ts m => MonadTransControlHandler tag ts m where

Associated Types

type StEffsT ts a

Methods

liftHandlerWith :: (Run (HandlerT tag ts) -> m a) -> HandlerT tag ts m a

restoreHandlerT :: m (StEffsT ts a) -> HandlerT tag ts m a

type family Handles (t :: HandlerK) (eff :: EffectK) :: Bool

An open type family that is used to determine which effects ought to be handled by which handlers. If Handles t eff ~ 'True for some handler t and effect eff, the handler will be used to handle any effects sent to it via send; otherwise, the effect will be lifted to the next handler in the stack.

It is important that Handles t is total in its argument; that is, it is important that effects that cannot be handled produce 'False, not just that effects that can be handled produce 'True. The == type family is provided for this purpose: If a handler only handles a single effect, its Handles instance should look like the following:

type Handles MyEffectT eff = eff == MyEffect

If it handles multiple effects, it can use the Elem type family instead:

type Handles MyEffectT eff = eff `Elem` '[MyEffect1, MyEffect2]

More complex Handles instances are possible, but not generally very useful.

type family a == b where ...

Boolean equality on types.

This is essentially the same as == from Data.Type.Equality, but the version from Data.Type.Equality is written in such a way that allows GHC to deduce more information from 'True results but causes trouble when trying to compute the equality of rigid type variables. This definition uses a simpler version.

Equations

a == a = 'True

_ == _ = 'False

type family Elem (x :: k) (xs :: [k]) :: Bool where ...

Checks if x is in the type-level list xs (like elem, but at the type level).

Equations

Elem _ '[] = 'False

Elem x (x ': _) = 'True

Elem x (_ ': xs) = Elem x xs

class Handle (handles :: Bool) eff t m where

Methods

handle :: (eff (t m) => t m a) -> ((MonadTransControl t, eff m) => t m a) -> EffT t m a

class (Monad m, Handle (Handles t eff) eff t m) => Send eff t m where

A typeclass used to lift effectful actions into effect handlers. This is not necessary to use directly when using effects, but it is used as part of defining new effects. Every effect should be given an instance on EffT of the shape

instance Send eff t m => eff (EffT t m) where

where eff is replaced by the actual effect in question. Each method should be implemented using send or sendWith: send for algebraic/first-order operations and sendWith for scoped/higher-order ones.

There is no need to define any additional instances of this class.

Minimal complete definition

sendWith

Methods

send :: (forall n. eff n => n a) -> EffT t m a

Constructs an EffT t m a computation for an algebraic/first-order operation. Each first-order method in the EffT instance for a given effect should have the shape

method a b c = send @eff (method a b c)

where method a b c should be replaced with the method and its arguments, and eff should be replaced with the type of the effect. The explicit type application is necessary because eff only appears in a constraint in the type signature for send, which GHC cannot automatically infer.

send @eff m is equivalent to sendWith @eff m (lift m).

sendWith

Arguments

:: (eff (t m) => t m a)
An action to run in the current handler.

-> ((MonadTransControl t, eff m) => t m a)
An action that delegates to the underlying monad.

-> EffT t m a

Constructs an EffT t m a computation for a higher-order/scoped effect eff from two actions:

A “run” action, which executes the effect in the (t m) monad given (t m) has an instance of eff.

A “lift” action, which lifts the effect through (t m) into m given that t has a MonadTransControl instance and m has an instance of eff.

Each higher-order method in the EffT instance for a given effect should use sendWith to specify how it ought to be lifted through effect handlers. For example, the definition of local looks like this:

local f m = sendWith @(Reader r)
  (local f (runEffT m))
  (controlT $ \lower -> local f (lower $ runEffT m))

With this instance in place, Reader r can automatically be used with EffT t m a. Transformers that can handle the Reader r effect (i.e. ones for which Handles t (Reader r) ~ 'True) will use their Reader r instances, while other transformers will delegate to the underlying monad.

controlT :: (MonadTransControl t, Monad m, Monad (t m)) => (Run t -> m (StT t a)) -> t m a

Using MonadTransControl, lifts a higher-order effectful operation into the underlying monad. It is named by analogy to control, since both are intended for lifting “control operations,” i.e. operations that affect control flow.

controlT f is equivalent to restoreT . pure =<< liftWith f, but it is rare that restoreT or liftWith need to be used directly.

type family All (cs :: [k -> Constraint]) (a :: k) :: Constraint where ...

Equations

All '[] _ = ()

All (c ': cs) a = (c a, All cs a)

class All effs m => Can (effs :: [EffectK]) m

A helper type for combining effect constraints: (Can '[e1, e2, e3, ..., en] m) is equivalent to (e1 m, e2 m, e3 m, ..., en m). The constraint Can effs m should be read as specifying that the monad m can perform all of the effects in the list effs.