Control.Effect.State

Documentation

data State s :: Effect where

The State s effect provides access to a single cell of mutable state of type s.

Constructors

Get :: State s m s

Put :: ~s -> State s m ()

get :: State s :< effs => Eff effs s

Retrieves the current value of the state.

put :: State s :< effs => s -> Eff effs ()

Replaces the current state with the given value.

modify :: State s :< effs => (s -> s) -> Eff effs ()

Modifies the current state by applying the given function to it.