Control.Effect.Writer

Documentation

data Writer w :: Effect where

The Writer w effect allows the accumulation of monoidal values of type w.

Instances should obey the following laws:

tell x *> tell y ≡ tell (x <> y)

listen (tell x) ≡ (x,) <$> tell x

censor f (tell x) ≡ tell (f x)

Constructors

Tell :: w -> Writer w m ()

Listen :: Eff (Writer w ': effs) a -> Writer w (Eff effs) (w, a)

Censor :: (w -> w) -> Eff (Writer w ': effs) a -> Writer w (Eff effs) a

tell :: Writer w :< effs => w -> Eff effs ()

Appends the given value to the current output.

listen :: Writer w :< effs => Eff (Writer w ': effs) a -> Eff effs (w, a)

Executes the given action and includes its output in the result.

censor :: Writer w :< effs => (w -> w) -> Eff (Writer w ': effs) a -> Eff effs a

Executes the given action and modifies its output by applying the given function.